chore: upgrade kafkaflow.sample.batchoperations to modern c# practices
