+++ samples/KafkaFlow.Sample.BatchOperations/SampleBatchMessage.cs
@@ -1,0 +1,6 @@
+namespace KafkaFlow.Sample.BatchOperations;
+
+internal class SampleBatchMessage
+{
+    public string Text { get; set; }
+}
